Unflatten agents directory and move drivers
This puts driver code close to the Agents. All imports are absolute imports.
