+++ ocs/agents/ocs_plugin_standard.py
@@ -10,11 +10,11 @@ root = os.path.abspath(os.path.split(__file__)[0])
 print(root)
 
 for n,f in [
-        ('RegistryAgent', 'registry.py'),
-        ('AggregatorAgent', 'aggregator_agent.py'),
-        ('HostManager', 'host_manager.py'),
-        ('FakeDataAgent', 'fake_data_agent.py'),
-        ('InfluxDBAgent', 'influxdb_publisher.py'),
-        ('BarebonesAgent', 'barebones_agent.py'),
+        ('RegistryAgent', 'registry/agent.py'),
+        ('AggregatorAgent', 'aggregator/agent.py'),
+        ('HostManager', 'host_manager/agent.py'),
+        ('FakeDataAgent', 'fake_data/agent.py'),
+        ('InfluxDBAgent', 'influxdb_publisher/agent.py'),
+        ('BarebonesAgent', 'barebones/agent.py'),
 ]:
     ocs.site_config.register_agent_class(n, os.path.join(root, f))
