+++ tests/agents/test_host_manager.py
@@ -1,4 +1,4 @@
-from ocs.agents.host_manager import HostManager
+from ocs.agents.host_manager.agent import HostManager
 
 import pytest_twisted
