+++ tests/integration/test_aggregator_agent_integration.py
@@ -16,8 +16,10 @@ pytest_plugins = ("docker_compose")
 
 wait_for_crossbar = create_crossbar_fixture()
 run_agent = create_agent_runner_fixture(
-    '../ocs/agents/aggregator_agent.py',
-    'aggregator-local')
+    '../ocs/agents/aggregator/agent.py',
+    'aggregator-local',
+    args=["--log-dir", "./logs/"],
+)
 client = create_client_fixture('aggregator-local')
